Handle media uploads and twitter better
When NOT using the media endpoint (that is, when getting a
form-submitted list of photos along with the content of a post), we need
to upload the files to the source directory. That was working, but I
was **not** checking that the necessary source directory was present, so
the actual uploads were failing.

Also, the Twitter media upload process returns an object. I only need
the string ID of the result if it succeeds.  This more gracefully
handles errors, insofar as it ignores them. :)