Better timezone handling; use file_put_contents
I run my PHP in a container, which defaults to UTC time; but I want US
Eastern time used by the `date()` function when building date-based
paths.  The way this previously worked was that I set the default
timezone **after** I'd used a `date()` function in `config.php`, which
resulted in a UTC-based date that was "tomorrow" from my local time
zone's perspective.

Explicitly setting the timezone in `config.php` first, rather than
making this a variable in the `$config` hash does what I want: sets the
default timezone as early as possible, and allows all subsequent
`date()` invocations to render US Eastern times.

This also streamlines the creation of the Markdown files by using
`file_put_contents`, which states:
```
This function is identical to calling fopen(), fwrite() and fclose()
successively to write data to a file.
```