Refine source interactions for replies/reposts/etc
Replies and reposts may largely interact with silos, but they need not,
so naming the functions related to those with `silo_` was misleading.

While the only implemented function for replies and reposts deals with
Twitter, this system should support connecting to **any** source URL
listed as a repost or reply or bookmark, and do something interesting.
That may be fetching all or some of the source content, or something
else.

TODO: need to add a generic handler for sources not explicitly defined.

This commit also adds support for bookmarking Tweets, which will add the
Tweet's author and contents to the front matter, exactly as we're doing
for reposts and replies.

Button up the `post_type_discovery` function to reduce some duplication
of code.